Graphs: Support floating-point scale factors
Add UAT support for storing and interpreting double-precision
floating point. Use that for the "Y Axis Scale Factor" in
graphs. This is forwards-compatible, but if a non-integer
number is stored in a UAT file and an older version of Wireshark
opens it, it will refuse to parse and fall back to the default
plots.

Fix #20667

+++ ui/qt/graph.cpp
@@ -85,7 +85,7 @@ void Graph::setVisible(bool visible)
     }
 }
 
-void Graph::setYAxisFactor(unsigned int y_axis_factor)
+void Graph::setYAxisFactor(double y_axis_factor)
 {
     y_axis_factor_ = y_axis_factor;
 }
