+++ ui/qt/plot_dialog.h
@@ -54,7 +54,7 @@ typedef struct _plot_settings_t {
     unsigned color;
     uint32_t style;
     char* yfield;
-    uint32_t y_axis_factor;
+    double y_axis_factor;
 } plot_settings_t;
 
 static const value_string graph_style_vs[] = {
@@ -112,7 +112,7 @@ private:
     int getLastPlotIdx();
     /* Add plot with all defined parameters. */
     void addPlot(bool checked, const QString& name, const QString& dfilter, QRgb color_idx,
-        Graph::PlotStyles style, const QString& yfield, int y_axis_factor = Graph::default_y_axis_factor_);
+        Graph::PlotStyles style, const QString& yfield, double y_axis_factor = Graph::default_y_axis_factor_);
     /* Add one of the two (four) default plots. */
     void addDefaultPlot(bool enabled, bool filtered);
